[UI] 코멘트 일기 작성시 하단 노티스 추가
