+++ presentation/src/main/java/com/movingmaker/presentation/util/BindingAdapter.kt
@@ -105,100 +105,6 @@ fun bindVisible(view: TextView, text: LiveData<String>) {
     }
 }
 
-@BindingAdapter("diaryType", "isExpand", "selectedDate")
-fun bindVisibleWithDiaryType(
-    view: View,
-    diaryType: LiveData<DIARY_TYPE>,
-    isExpand: LiveData<Boolean>,
-    tempSelectedDiaryDate: LiveData<String>
-) {
-    when (isExpand.value) {
-        true -> {
-            when (view.id) {
-                R.id.sendButton, R.id.saveButton -> {
-                    view.visibility = View.INVISIBLE
-                }
-                R.id.selectAloneDiaryInRadioButton -> {
-                    when (diaryType.value) {
-                        DIARY_TYPE.COMMENT_DIARY -> {
-                            view.visibility = View.GONE
-                        }
-                        DIARY_TYPE.ALONE_DIARY -> {
-                            view.visibility = View.VISIBLE
-                        }
-                        else -> {}
-                    }
-                }
-                R.id.selectCommentDiaryInRadioButton -> {
-                    when (diaryType.value) {
-                        DIARY_TYPE.COMMENT_DIARY -> {
-                            view.visibility = View.VISIBLE
-                        }
-                        DIARY_TYPE.ALONE_DIARY -> {
-                            view.visibility = View.GONE
-                        }
-                        else -> {}
-                    }
-                }
-                else -> {}
-            }
-        }
-        else -> {
-            when (view.id) {
-                R.id.sendButton -> {
-                    when (diaryType.value) {
-                        DIARY_TYPE.COMMENT_DIARY -> {
-                            view.visibility = View.VISIBLE
-                        }
-                        DIARY_TYPE.ALONE_DIARY -> {
-                            view.visibility = View.INVISIBLE
-                        }
-                        else -> {}
-                    }
-                }
-                R.id.saveButton -> {
-                    when (diaryType.value) {
-                        DIARY_TYPE.COMMENT_DIARY -> {
-                            view.visibility = View.INVISIBLE
-                        }
-                        DIARY_TYPE.ALONE_DIARY -> {
-                            view.visibility = View.VISIBLE
-                        }
-                        else -> {}
-                    }
-                }
-                R.id.selectCommentDiaryInRadioButton, R.id.selectAloneDiaryInRadioButton -> {
-                    view.visibility = View.GONE
-                }
-                else -> {
-                    (view as TextView)
-                    tempSelectedDiaryDate.value?.let {
-                        val selectedDiaryDate = ymdToDate(it)
-                        selectedDiaryDate?.let { selectedDate ->
-                            if (selectedDate < getCodaToday()) {
-                                view.isEnabled = false
-                                view.setCompoundDrawablesWithIntrinsicBounds(0, 0, 0, 0)
-                            } else {
-                                view.isEnabled = true
-                                view.setCompoundDrawablesWithIntrinsicBounds(
-                                    0,
-                                    0,
-                                    R.drawable.ic_arrow_down,
-                                    0
-                                )
-                            }
-                        }
-                    }
-                    view.text = when (diaryType.value) {
-                        DIARY_TYPE.COMMENT_DIARY -> view.context.getString(R.string.diary_type_comment)
-                        else -> view.context.getString(R.string.diary_type_alone)
-                    }
-                }
-            }
-        }
-    }
-}
-
 @BindingAdapter("nonClickableActivity", "nonClickableLoading")
 fun bindNonClickableLoading(view: ProgressBar, activity: Activity, state: LiveData<Boolean>) {
     when (state.value) {
