+++ presentation/src/main/res/layout/fragment_mydiary_writediary.xml
@@ -200,6 +200,7 @@
             android:layout_height="22dp"
             android:layout_marginTop="11dp"
             android:layout_marginEnd="18dp"
+            app:diaryType="@{vm.selectedDiaryType}"
             android:text="@{@string/comment_diary_text_limit(vm.diaryContent.length)}"
             android:textColor="@color/text_brown"
             app:layout_constraintEnd_toEndOf="parent"
